Add files via upload
Assignment Description:

Question#2.  Write a static method inList() that performs an iterative sequential search of a list by using only operations of the ADT list.  The method should return true if a given item is in a given list.

Write a main() to use this method inList().
Use the array from figure 19-2, p543.  a[] = { 9, 5, 8, 4, 7 }
Search for 8 (return true)  -and-  6 (return false)